[Factory Reset] Fix keeping settings (#4263)

+++ src/src/ESPEasyCore/ESPEasyWifi.cpp
@@ -726,7 +726,7 @@ void SetWiFiTXpower(float dBm, float rssi) {
   int8_t power = dBm * 4;
   if (esp_wifi_set_max_tx_power(power) == ESP_OK)  {
     if (esp_wifi_get_max_tx_power(&power) == ESP_OK)  {
-      dBm = power / 4;
+      dBm = static_cast<float>(power) / 4.0f;
     }
   }
   #endif
